+++ curifactory/experiment.py
@@ -176,7 +176,7 @@ def run_experiment(  # noqa: C901 -- TODO: this does need to be broken up at som
         if parallel is not None:
             run_string += f" --parallel {parallel}"
         if parallel_mode:
-            run_string += " --parallel-mode"
+            run_string += " --parallel-safe"
         if global_args_indices is not None:
             for index_range in global_args_indices:
                 run_string += f" --global-indices {index_range}"
@@ -1126,11 +1126,13 @@ Examples:
     # ---- PARAMETERS ----
     parameters_group.add_argument(
         "-p",
+        "--params",
         dest="parameters_name",
         action="append",
         help="The name of a parameters python file. Does not need to include path or .py. You can specify multiple -p arguments to run all parameters",
     ).completer = params_completer
     parameters_group.add_argument(
+        "-n",
         "--names",
         dest="args_name",
         action="append",
@@ -1276,7 +1278,7 @@ Examples:
         help="Divide the parameters among n subprocesses in order to cache intermediate results in parallel, and then re-run in original process with cached data.",
     )
     parallel_group.add_argument(
-        "--parallel-mode",
+        "--parallel-safe",
         dest="parallel_mode",
         action="store_true",
         help="Runs an experiment and caches intermediate values but does not report or touch experiment store and parameter registry. Note that using this flag manually is not equivalent to running an experiment with the --parallel flag, as file locks don't get used when calling this manually, potentially creating race conditions.",